--- src/io/fasta/mod.rs
@@ -1,5 +1,0 @@
-pub mod reader;
-
-use super::*;
-
-const SEQ_START: u8 = b'>';
