--- src/io/fasta/reader/mod.rs
@@ -1,3 +1,0 @@
-pub mod fai;
-
-use super::*;
