--- src/io/fastq/mod.rs
@@ -1,6 +1,0 @@
-pub mod reader;
-
-use super::*;
-
-const SEQ_START: u8 = b'@';
-const PLUS: u8 = b'+';
