+++ src/io/mod.rs
@@ -1,5 +1,3 @@
 pub mod fai;
-pub mod fasta;
-pub mod fastq;
 
 mod common;
